+++ lib/grid_codec/types/i16.ex
@@ -138,6 +138,11 @@ defmodule GridCodec.Types.I16 do
     if value == @null_val, do: nil, else: value
   end
 
+  @impl true
+  def validate_ast(var, field, mod) do
+    GridCodec.Types.Integer.gen_signed_validate_ast(var, field, mod, 16, :i16)
+  end
+
   if Code.ensure_loaded?(GridCodec.Generators) do
     @impl true
     def generator, do: GridCodec.Generators.i16()