+++ lib/grid_codec/types/i64.ex
@@ -143,6 +143,11 @@ defmodule GridCodec.Types.I64 do
     end
   end
 
+  @impl true
+  def validate_ast(var, field, mod) do
+    GridCodec.Types.Integer.gen_signed_validate_ast(var, field, mod, 64, :i64)
+  end
+
   if Code.ensure_loaded?(GridCodec.Generators) do
     @impl true
     def generator, do: GridCodec.Generators.i64()