+++ lib/grid_codec/types/i8.ex
@@ -95,6 +95,11 @@ defmodule GridCodec.Types.I8 do
     if value == @null_val, do: nil, else: value
   end
 
+  @impl true
+  def validate_ast(var, field, mod) do
+    GridCodec.Types.Integer.gen_signed_validate_ast(var, field, mod, 8, :i8)
+  end
+
   if Code.ensure_loaded?(GridCodec.Generators) do
     @impl true
     def generator, do: GridCodec.Generators.i8()