+++ lib/grid_codec/types/u16.ex
@@ -140,6 +140,11 @@ defmodule GridCodec.Types.U16 do
     end
   end
 
+  @impl true
+  def validate_ast(var, field, mod) do
+    GridCodec.Types.Integer.gen_unsigned_validate_ast(var, field, mod, 16, :u16)
+  end
+
   if Code.ensure_loaded?(GridCodec.Generators) do
     @impl true
     def generator, do: GridCodec.Generators.u16()