+++ lib/grid_codec/types/u64.ex
@@ -157,6 +157,11 @@ defmodule GridCodec.Types.U64 do
     end
   end
 
+  @impl true
+  def validate_ast(var, field, mod) do
+    GridCodec.Types.Integer.gen_unsigned_validate_ast(var, field, mod, 64, :u64)
+  end
+
   if Code.ensure_loaded?(GridCodec.Generators) do
     @impl true
     def generator, do: GridCodec.Generators.u64()